+++ lib/src/analyzers/lint_analyzer/rules/base_visitors/intl_base_visitor.dart
@@ -33,6 +33,7 @@ abstract class IntlBaseVisitor extends GeneralizingAstVisitor<void> {
 
   @override
   void visitFieldDeclaration(FieldDeclaration node) {
+    // ignore: deprecated_member_use
     if (node.fields.type?.as<TypeName>()?.name.name != 'String') {
       return;
     }