+++ lib/src/analyzers/lint_analyzer/rules/rules_list/avoid_returning_widgets/visitor.dart
@@ -36,6 +36,7 @@ class _Visitor extends RecursiveAstVisitor<void> {
 
   @override
   void visitClassDeclaration(ClassDeclaration node) {
+    // ignore: deprecated_member_use
     final classType = node.extendsClause?.superclass.type;
     if (!isWidgetOrSubclass(classType) && !isWidgetStateOrSubclass(classType)) {
       return;