+++ lib/src/analyzers/lint_analyzer/rules/rules_list/prefer_single_widget_per_file/visitor.dart
@@ -15,6 +15,7 @@ class _Visitor extends SimpleAstVisitor<void> {
   void visitClassDeclaration(ClassDeclaration node) {
     super.visitClassDeclaration(node);
 
+    // ignore: deprecated_member_use
     final classType = node.extendsClause?.superclass.type;
     if (isWidgetOrSubclass(classType) &&
         (!_ignorePrivateWidgets || !Identifier.isPrivateName(node.name.name))) {