+++ lib/src/analyzers/lint_analyzer/metrics/metrics_factory.dart
@@ -33,10 +33,14 @@ final _implementedMetrics = <String, Metric Function(Map<String, Object>)>{
 
 Iterable<Metric> getMetrics({
   required Map<String, Object> config,
+  Iterable<String> patternsDependencies = const [],
   EntityType? measuredType,
 }) {
-  final _metrics =
-      _implementedMetrics.keys.map((id) => _implementedMetrics[id]!(config));
+  final _metrics = _implementedMetrics.keys.map(
+    (id) => _implementedMetrics[id]!(
+      !patternsDependencies.contains(id) ? config : {},
+    ),
+  );
 
   return measuredType != null
       ? _metrics