+++ test/resources/number_of_parameters_metric_example.dart
@@ -45,3 +45,27 @@ class NumberOfParametersMetric extends FunctionMetric<int> {
     return MetricComputationResult(value: parametersCount ?? 0);
   }
 }
+
+class Foo {
+  final String a;
+  final String b;
+  final String c;
+  final String d;
+  final String e;
+
+  const Foo({
+    required this.a,
+    required this.b,
+    required this.c,
+    required this.d,
+    required this.e,
+  });
+
+  Foo copyWith({String a, String b, String c, String d, String e}) => Foo(
+        a: a ?? this.a,
+        b: b ?? this.b,
+        c: c ?? this.c,
+        d: d ?? this.d,
+        e: e ?? this.e,
+      );
+}