+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * feat: improve static code diagnostic `prefer-extracting-callbacks`: don't trigger on empty function blocks and ignore Flutter builder functions.
 * feat: improve unused files check, add support for `vm:entry-point` annotation.
 * fix: compute NumberOfParametersMetric only for functions and methods.
+* fix: skip synthetic tokens while compute 'Source lines of Code'.
 
 ## 4.3.3